52_minotaur: Updated all tests to work without a database connection
Merge branch '52_minotaur' into 'main'

Reviewed by 
Approved by Approved-by: Rasmus Jakobsson <c20rjn@cs.umu.se>

Description
https://trello.com/c/r6KvQh6f

The backend tests were not mocking the database correctly since it was using @SpringBootTest annotation. This annotation still tries to create a connection to the database in the .yml files. I changed all tests to work even with a faulty database connection since that is not the purpose of the tests.

Branch commits
* Updated all tests to work without a database connection