67_Phoenix: Added method to fetch a limited amount of excercises.
Merge branch '67_Phoenix' into 'main'

Reviewed by 
Approved by 

Description
https://trello.com/c/vQgZwjXT/67-implementation-av-%C3%B6vningar-vy

A simple method has been added to fetch a limited amount of exercises. This is needed for the frontend so that all exercises isn't fetched at once, we need this merged so that we can try it in the frontend.

Branch commits
* Added a method to fetch a limited amount of exercises.