67_Phoenix: Modified the method that gets a limited amount of exercises 2.0
Merge branch '67_Phoenix' into 'main'

Reviewed by 
Approved by Approved-by: Erik Risberg <c18erg@cs.umu.se>

Description
https://trello.com/c/vQgZwjXT/67-implementation-av-%C3%B6vningar-vy

A simple method has been modified to fetch a limited amount of exercises. This is needed for the frontend so that all exercises isn't fetched at once, we need this merged so that we can try it in the frontend.

Branch commits
* Modified the method that gets a limited amount of exercises.