67_Phoenix: Added methods to fetch exercises just for testing/debugging purposes.
Merge branch '67_Phoenix' into 'main'

Reviewed by 
Approved by Approved-by: Erik Risberg <c18erg@cs.umu.se>

Description
https://trello.com/c/vQgZwjXT/67-implementation-av-%C3%B6vningar-vy

Simple methods has been added and modified to fetch a limited amount of exercises. This is needed for the frontend so that all exercises isn't fetched at once, we need this merged so that we can try it in the frontend. Used for debugging/testing.

Branch commits
* Added methods to fetch exercises just for testing purposes.