71_minotaur: AddUserComponent in workouts
Merge branch '71_minotaur' into 'main'

Reviewed by Reviewed-by: Emmy Lindgren <id19eln@cs.umu.se>
Approved by Approved-by: Emmy Lindgren <id19eln@cs.umu.se>

Description
A new 'Add user' component for the 'new workout' page has been
created. This component is responsible for searching for
users and displaying them in a list. It also keeps track of
selected users, which is retrieveable using the hook.

A new common component 'searchable dropdown' has also been
added.

In the future, the search-logic should be moved to the backend.

This patch is **NOT** done, and needs future tests.

Branch commits
* 71_minotaur: AddUserComponent in workouts

A new 'Add user' component for the 'new workout' page has been
created. This component is responsible for searching for
users and displaying them in a list. It also keeps track of
selected users, which is retrieveable using the hook.

In the future, the search-logic should be moved to the backend.