34 kraken: Created the addActivity popup and some adjustments to styling
Merge branch '34_kraken' into 'main'

Reviewed by 
Approved by 

Description
34_kraken:
https://trello.com/c/EIcMiK5z

33_kraken:
https://trello.com/c/t0nRwK7P

Since both these tickets handled the same popu it was combined to a single merge. Everything is contained in a single file.

Some changes to css in the _TechniqueCard_ component were made to look more like the component in Figma. The button on the edit/create workout must be implemented to show this popu. The "continue" button on this component need to be linked to the other page in later merges.

Bugfixing and testing is needed but done at a later date.

Branch commits
* Removed test suite for add activity

* Further work on the addActivity popup and some adjustments to css

* Changed some css and functionality for the add activity popup

* reverted changes made for testing purposes

* Done with overall layout and functionallity for addActivity

* Started with add technique or exericse to workout popup

* reverted changes made for testing purposes

* Done with overall layout and functionallity for addActivity

* add activity progress

* mob programming results

* Start to thing

* Started with add technique or exericse to workout popup