120_chimera: Fix the cookie expiration bug
Merge branch '120_chimera' into 'main'

Reviewed by Reviewed-by: Carl Gulliksson <dv21cgn@cs.umu.se>
Approved by Approved-by: Carl Gulliksson <dv21cgn@cs.umu.se>

Description
Fixes ticket [#120](https://trello.com/c/4Lw1QfnW/120-cookiebuggen)

The reason why the page turned white was because the cookie expired, and the user
was not redirected to the login page. 

There was a mechanism implemented for notifying the user that the cookie will expire, however, 
the implementation was flawed. It always used the same timeout instead of calculating 
the delta for the cookie expiration, hence it was never shown.

The issue is now fixed, and the user will be redirected to the login page when the cookie expires,
eliminating any cookie issues.

A future improvement might be to allow the cookie to be refreshed, however, there is currently no support
in the backend for refreshing tokens.

Branch commits
* Add warning when session is about to expire

* Null check

* Fix the cookie expiration bug