135 _medusa: Standardisering av felmeddelanden vid input
Merge branch '135_medusa' into 'main'

Reviewed by 
Approved by Approved-by: Alexander Lindblom <c20alm@cs.umu.se>
Approved-by: Christoffer Billman <id19cbn@cs.umu.se>

Description
[#135](https://trello.com/c/UkLnKSsa/135-standardisering-av-felmeddelanden-vid-input)
The input now takes a prop  "errorMessage". When this prop is set to a non-empty string, it will show the contents of the string as an error message, and change the style of the input.

Branch commits
* Wrote some tests for InputTextField and TextArea

* Fixed some styling issue.

* styling fix: Fixed height on input fields.
Changed from 100% to 6 and 4 rem.

* Updates TextArea to take prop errorMessage.

* converted css to css module.

* Added error styles and prop for input.

* Changed CSS to use to CSS modules.