158_medusa: Added min and max date to date picker
Merge branch '158_medusa' into 'main'

Reviewed by 
Approved by Approved-by: André Marklund <ens19amd@cs.umu.se>

Description
https://trello.com/c/A0SVV9Zk/158-just-nu-g%C3%A5r-det-att-v%C3%A4lja-datum-fr%C3%A5n-och-till-d%C3%A4r-fr%C3%A5n-%C3%A4r-efter-till-fixa-detta-p%C3%A5-pass-sidan

Also implemented this in workoutIndex. There is a small bug that you can manually type a from date that is larger that the to date, but it is only possible when the first three digits of the date is the same. Do not know how to fix this

Branch commits
* fixed so from date can't be after to date and vice versa

* Added min and max date