114_minotaur: Moved systests to infra
Merge branch '114_minotaur' into 'main'

Reviewed by 
Approved by Approved-by: Alexander Arvidsson Örnberg <dv21aag@cs.umu.se>

Description
System tests have been moved to infra for importing into the other
repositories. There are two scripts: `run-systest.sh` and
`find-version.sh`. `run-systest.sh` takes all versions and a runner
id, and runs the systemtests. `find-version.sh` takes a system
(db, frontend, api), and the commit message, and uses that
to parse out the version.

For this to work, we need to publish a `:staging-<feature branch>`
tag to the docker repository. This could become a problem, as
dockerhub would fill up. We have considered other solutions
but this seems the simplest, but should be fixed in the future.

System tests are also run when changes to the database has been made.

Branch commits
* 114_minotaur: Moved systests to infra

System tests have been moved to infra for importing into the other
repositories. There are two scripts: `run-systest.sh` and
`find-version.sh`. `run-systest.sh` takes all versions and a runner
id, and runs the systemtests. `find-version.sh` takes a system
(db, frontend, api), and the commit message, and uses that
to parse out the version.

For this to work, we need to publish a `:staging-<feature branch>`
tag to the docker repository. This could become a problem, as
dockerhub would fill up. We have considered other solutions
but this seems the simplest, but should be fixed in the future.

System tests are also run when changes to the database has been made.